refactor(frontend): split desktop session workflow responsibilities
