+++ apps/desktop/src/hooks/sessions/lifecycle/use-session-history-hydration.ts
@@ -1,51 +1,35 @@
-import {
-  type SessionEventEnvelope,
-  type TranscriptState,
-} from "@anyharness/sdk";
 import { useCallback } from "react";
 import {
   mergeFetchedHistoryWithExistingEvents,
   mergeFetchedHistoryWithNewerEvents,
 } from "@/lib/domain/sessions/history/history-event-merge";
-import { resolveSessionStatus } from "@proliferate/product-domain/sessions/activity";
 import {
   appendHistoryTail,
   replaySessionHistory,
 } from "@/lib/domain/sessions/stream/stream-state";
 import { logLatency } from "@/lib/infra/measurement/debug-latency";
 import {
-  finishOrCancelMeasurementOperation,
-  markOperationForNextCommit,
   recordMeasurementMetric,
   recordMeasurementWorkflowStep,
   startMeasurementOperation,
 } from "@/lib/infra/measurement/debug-measurement";
 import type {
   MeasurementOperationId,
-  MeasurementOperationKind,
-  MeasurementSurface,
 } from "@/lib/domain/telemetry/debug-measurement-catalog";
 import { uniqueMeasurementOperationIds } from "@/lib/infra/measurement/operation-ids";
-import { batchSessionStoreWrites } from "@/lib/infra/scheduling/react-batching";
-import { scheduleAfterNextPaint } from "@/lib/infra/scheduling/schedule-after-next-paint";
-import { fetchSessionHistory } from "@/lib/workflows/sessions/session-runtime";
-import { activityFromTranscript } from "@/lib/domain/sessions/directory/directory-activity";
+import { fetchSessionHistory } from "@/lib/access/anyharness/session-runtime";
 import { useLinkedSessionMounting } from "@/hooks/chat/workflows/subagents/use-linked-session-mounting";
-import { useSessionDirectoryStore } from "@/stores/sessions/session-directory-store";
 import { getSessionRecord } from "@/stores/sessions/session-records";
-import { useSessionTranscriptStore } from "@/stores/sessions/session-transcript-store";
-import type { SessionRuntimeRecord } from "@/stores/sessions/session-types";
-
-const SESSION_APPLY_MEASUREMENT_SURFACES: readonly MeasurementSurface[] = [
-  "session-transcript-pane",
-  "transcript-list",
-  "chat-surface",
-  "header-tabs",
-  "workspace-sidebar",
-  "global-header",
-  "chat-composer-dock",
-];
-const SESSION_HISTORY_APPLY_MAX_DURATION_MS = 30_000;
+import {
+  applyHistoryStateToStores,
+  finishStandaloneApplyOperation,
+  isSessionHistoryTimeoutAbort,
+  markSessionApplyForNextCommit,
+  recordHistoryStateCounts,
+  resolveHistoryApplyOperationKind,
+  SESSION_APPLY_MEASUREMENT_SURFACES,
+  SESSION_HISTORY_APPLY_MAX_DURATION_MS,
+} from "@/hooks/sessions/lifecycle/session-history-hydration-helpers";
 
 interface SessionHistoryHydrationOptions {
   afterSeq?: number;
@@ -397,105 +381,4 @@ export function useSessionHistoryHydration() {
   return {
     rehydrateSessionSlotFromHistory,
   };
-}
-
-function isSessionHistoryTimeoutAbort(error: unknown): boolean {
-  return error instanceof Error
-    && error.name === "AbortError"
-    && error.message === "Session history request timed out";
-}
-
-function resolveHistoryApplyOperationKind(input: {
-  afterSeq?: number;
-  beforeSeq?: number;
-}): MeasurementOperationKind {
-  if (input.beforeSeq != null) {
-    return "session_history_older_chunk";
-  }
-  if (input.afterSeq != null) {
-    return "session_history_tail_reconcile";
-  }
-  return "session_history_initial_hydrate";
-}
-
-function finishStandaloneApplyOperation(
-  operationId: MeasurementOperationId | null,
-  reason: "completed" | "aborted" | "error_sanitized",
-  waitForPaint = false,
-): void {
-  if (!operationId) {
-    return;
-  }
-  const finish = () => finishOrCancelMeasurementOperation(operationId, reason);
-  if (waitForPaint) {
-    scheduleAfterNextPaint(finish);
-    return;
-  }
-  finish();
-}
-
-function markSessionApplyForNextCommit(operationId: MeasurementOperationId | null | undefined): void {
-  if (!operationId) {
-    return;
-  }
-  markOperationForNextCommit(operationId, SESSION_APPLY_MEASUREMENT_SURFACES);
-}
-
-function applyHistoryStateToStores(
-  sessionId: string,
-  currentRecord: SessionRuntimeRecord,
-  nextState: {
-    events: SessionEventEnvelope[];
-    transcript: TranscriptState;
-  },
-): void {
-  const status = resolveSessionStatus(currentRecord.status, {
-    executionSummary: currentRecord.executionSummary,
-    streamConnectionState: currentRecord.streamConnectionState,
-    transcript: nextState.transcript,
-  });
-  batchSessionStoreWrites(() => {
-    useSessionTranscriptStore.getState().patchEntry(sessionId, {
-      events: nextState.events,
-      transcript: nextState.transcript,
-    });
-    useSessionDirectoryStore.getState().patchEntry(sessionId, {
-      status,
-      modeId: nextState.transcript.currentModeId ?? currentRecord.modeId,
-      activity: activityFromTranscript(nextState.transcript, {
-        status,
-        executionSummary: currentRecord.executionSummary,
-      }),
-    });
-  });
-}
-
-function recordHistoryStateCounts(
-  operationId: MeasurementOperationId | null | undefined,
-  phase: "before" | "after",
-  events: readonly SessionEventEnvelope[],
-  transcript: TranscriptState,
-): void {
-  if (!operationId) {
-    return;
-  }
-  const isBefore = phase === "before";
-  recordMeasurementMetric({
-    type: "state_count",
-    operationId,
-    target: isBefore ? "session.history.events_before" : "session.history.events_after",
-    count: events.length,
-  });
-  recordMeasurementMetric({
-    type: "state_count",
-    operationId,
-    target: isBefore ? "session.history.turns_before" : "session.history.turns_after",
-    count: transcript.turnOrder.length,
-  });
-  recordMeasurementMetric({
-    type: "state_count",
-    operationId,
-    target: isBefore ? "session.history.items_before" : "session.history.items_after",
-    count: Object.keys(transcript.itemsById).length,
-  });
 }
