+++ apps/desktop/src/hooks/sessions/workflows/session-selection-measurement.ts
@@ -1,0 +1,21 @@
+export const SESSION_SWITCH_MEASUREMENT_SURFACES = [
+  "workspace-shell",
+  "chat-surface",
+  "session-transcript-pane",
+  "transcript-list",
+  "header-tabs",
+  "workspace-sidebar",
+] as const;
+
+export const SESSION_HOT_SWITCH_MEASUREMENT_SURFACES = [
+  "workspace-shell",
+  "chat-surface",
+  "session-transcript-pane",
+  "transcript-list",
+  "transcript-context-providers",
+  "transcript-row-list-router",
+  "transcript-virtualized-viewport",
+  "transcript-full-list",
+  "header-tabs",
+  "workspace-sidebar",
+] as const;
